feat: Add initial_import option
If initial_import is set to true, the godot project will be opened in a headless editor once, before exporting the binaries.
This triggers importing all assets and creating the `.godot` directory.
It is necessary for headless exports to work, if no `.godot` folder exists in the git repository.
The `.godot` folder is excluded from git in the default godot 4 `.gitignore`.
The import is only executed, if `use_godot_4` is true, since godot 3 has dedicated headless builds and does not support the `--headless` argument.